fix: fix bug where email was not sending. update: refactor the webhooks processor to be used with adonisjs-scheduler package. update: pass transaction to tierService methods - calcEffectiveTier fix: register Webhook and Confirmation email services.

+++ database/migrations/1760904323858_create_tier_audit_logs_table.ts
@@ -7,7 +7,7 @@ export default class extends BaseSchema {
     this.schema.createTable(this.tableName, (table) => {
       table.bigIncrements('id').primary()
       table
-        .bigInteger('user_id')
+        .integer('user_id')
         .unsigned()
         .notNullable()
         .references('id')
